feat: 프롬프트 상세 페이지 접근 시 서버에서 데이터 Prefetch
